requirements.txt
Flask==0.12.2
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==4.2.1
RPi.GPIO==0.6.2
Werkzeug==0.12.2
chardet==2.3.0
click==6.7
colorama==0.3.2
gpiozero==1.2.0
html5lib==0.999
itsdangerous==0.24
mcpi==0.1.1
numpy==1.8.2
olefile==0.44
pgzero==1.1
picamera==1.10
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pigpio==1.30
pygame==1.9.2a0
pyserial==2.6
python-apt==0.9.3.12
python-debian==0.1.27
requests==2.4.3
six==1.8.0
urllib3==1.9.1
wheel==0.24.0
